requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
autobahn==24.4.2
Automat==24.8.1
av==12.0.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
class-registry==2.1.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.0
cryptography==43.0.3
cycler==0.12.1
daphne==4.1.2
distlib==0.3.8
Django==5.1.2
django-crispy-forms==2.3
django-filter==24.3
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djitellopy==2.5.0
dlib==19.24.2
dm-tree==0.1.8
drf-yasg==1.21.8
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.3
filters==1.3.2
Flask==3.0.3
flatbuffers==24.3.7
fonttools==4.50.0
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
gradio==4.24.0
gradio_client==0.14.0
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.1
hyperlink==21.0.0
idna==3.6
importlib_resources==6.4.0
imutils==0.5.4
incremental==24.7.2
inflection==0.5.1
itsdangerous==2.2.0
jax==0.4.25
jaxlib==0.4.25
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==3.0.5
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mediapipe==0.10.11
ml-dtypes==0.3.2
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.0
mtcnn==0.1.1
namex==0.0.7
networkx==3.2.1
nltk==3.9.1
numpy==1.26.2
opencv-contrib-python==4.9.0.80
opencv-python==4.8.1.78
opt-einsum==3.3.0
orjson==3.10.0
packaging==23.2
pandas==2.2.1
Pillow==10.1.0
platformdirs==4.2.0
plotly==5.18.0
protobuf==3.20.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycfg==0.1
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.17.2
PyJWT==2.9.0
PyMsgBox==1.0.9
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-Quartz==10.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytweening==1.0.7
pytz==2024.1
PyYAML==6.0.1
redis==5.1.1
referencing==0.34.0
regex==2024.7.24
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rubicon-objc==0.4.7
ruff==0.3.4
scipy==1.12.0
semantic-version==2.10.0
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
sqlparse==0.5.1
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.2
Twisted==24.7.0
txaio==23.1.1
typer==0.11.0
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.25.1
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.16.0
zope.interface==7.1.0
